Add timestamp to all reports
Currently, the `timestamp` field is only added to reports that are
processed with a Spartakus collector. If you run the volunteer with
`--database=stdout` or even `--database=bigquery://...` your records
will have no timestamp. This commit adds the timestamp field to reports
at the time of creation in the volunteer.