report: add cloud provider name to report
This commit modifies the Node struct and adds a new field
`CloudProvider`. This value is determined by taking the `<ProviderName>`
portion of the `ProviderID` reported in the node spec. Making this
change will require updating the database schema. The cloud provider
name is not PII and will provide useful insight into who is using
kubernetes and how.